+++ src/App.tsx
@@ -1,25 +1,25 @@
-import { Route, Routes } from "react-router-dom";
 import "./App.css";
+import { Route, Routes } from "react-router-dom";
+import { RecoilRoot } from "recoil";
 import Login from "./pages/Login";
 import Home from "./pages/Home";
 import SignUp from "./pages/Signup";
-import { RecoilRoot } from "recoil";
 import Layout from "./layout";
-import Test01Menu from "./pages/menus/Test01Menu";
-import Test02Menu from "./pages/menus/Test02Menu";
 import MyPage from "./pages/MyPage";
-import Test03Menu from "./pages/menus/Test03Menu";
-import Test04Menu from "./pages/menus/Test04Menu";
-import Test05Menu from "./pages/menus/Test05Menu";
-import CrosswordQuiz from "./pages/tests/test05/CrosswordQuiz";
+import Test01Menu from "./pages/tests/test01/Test01Menu";
 import Test01Level from "./pages/tests/test01/Test01Level";
 import Test01ScreenWrapper from "./pages/tests/test01/Test01ScreenWrapper";
+import Test02Menu from "./pages/tests/test02/Test02Menu";
 import Test02Level from "./pages/tests/test02/Test02Level";
 import Test02ScreenWrapper from "./pages/tests/test02/Test02ScreenWrapper";
+import Test03Menu from "./pages/tests/test03/Test03Menu";
 import Test03Level from "./pages/tests/test03/Test03Level";
 import Test03ScreenWrapper from "./pages/tests/test03/Test03ScreenWrapper";
+import Test04Menu from "./pages/tests/test04/Test04Menu";
 import Test04Level from "./pages/tests/test04/Test04Level";
 import Test04ScreenWrapper from "./pages/tests/test04/Test04ScreenWrapper";
+import Test05Menu from "./pages/tests/test05/Test05Menu";
+import CrosswordQuiz from "./pages/tests/test05/CrosswordQuiz";
 
 function LayoutRoutes() {
   return (
