+++ src/layout.tsx
@@ -1,4 +1,4 @@
-import Header from "./components/Header";
+import Header from "./components/common/Header";
 
 export default function Layout({ children }: any) {
   return (
