+++ src/pages/tests/test02/Test02Menu.tsx
@@ -1,4 +1,4 @@
-import Menus from "../../components/Menus";
+import Menus from "../../../components/tests/Menus";
 
 export default function Test02Menu() {
   return (
