+++ src/pages/tests/test04/Test04Menu.tsx
@@ -1,4 +1,4 @@
-import Menus from "../../components/Menus";
+import Menus from "../../../components/tests/Menus";
 
 export default function Test04Menu() {
   return (
